Refactor dotnet/nuget/binlog detectors to share as much as possible
