+++ src/Microsoft.ComponentDetection.Detectors/dotnet/DotNetComponentDetector.cs
@@ -1,16 +1,8 @@
 namespace Microsoft.ComponentDetection.Detectors.DotNet;
 
-using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-using System.IO;
-using System.Linq;
-using System.Reflection.PortableExecutable;
-using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
-using global::NuGet.Frameworks;
 using global::NuGet.ProjectModel;
 using Microsoft.ComponentDetection.Contracts;
 using Microsoft.ComponentDetection.Contracts.Internal;
@@ -19,18 +11,8 @@ using Microsoft.Extensions.Logging;
 
 public class DotNetComponentDetector : FileComponentDetector
 {
-    private const string GlobalJsonFileName = "global.json";
-    private readonly ICommandLineInvocationService commandLineInvocationService;
-    private readonly IDirectoryUtilityService directoryUtilityService;
-    private readonly IFileUtilityService fileUtilityService;
-    private readonly IPathUtilityService pathUtilityService;
+    private readonly DotNetProjectInfoProvider projectInfoProvider;
     private readonly LockFileFormat lockFileFormat = new();
-    private readonly ConcurrentDictionary<string, string?> sdkVersionCache = [];
-    private readonly JsonDocumentOptions jsonDocumentOptions =
-        new() { CommentHandling = JsonCommentHandling.Skip, AllowTrailingCommas = true };
-
-    private string? sourceDirectory;
-    private string? sourceFileRootDirectory;
 
     public DotNetComponentDetector(
         IComponentStreamEnumerableFactory componentStreamEnumerableFactory,
@@ -41,13 +23,15 @@ public class DotNetComponentDetector : FileComponentDetector
         IObservableDirectoryWalkerFactory walkerFactory,
         ILogger<DotNetComponentDetector> logger)
     {
-        this.commandLineInvocationService = commandLineInvocationService;
         this.ComponentStreamEnumerableFactory = componentStreamEnumerableFactory;
-        this.directoryUtilityService = directoryUtilityService;
-        this.fileUtilityService = fileUtilityService;
-        this.pathUtilityService = pathUtilityService;
         this.Scanner = walkerFactory;
         this.Logger = logger;
+        this.projectInfoProvider = new DotNetProjectInfoProvider(
+            commandLineInvocationService,
+            directoryUtilityService,
+            fileUtilityService,
+            pathUtilityService,
+            logger);
     }
 
     public override string Id => "DotNet";
@@ -60,93 +44,9 @@ public class DotNetComponentDetector : FileComponentDetector
 
     public override IEnumerable<string> Categories => ["DotNet"];
 
-    private static string TrimAllEndingDirectorySeparators(string path)
-    {
-        string last;
-
-        do
-        {
-            last = path;
-            path = Path.TrimEndingDirectorySeparator(last);
-        }
-        while (!ReferenceEquals(last, path));
-
-        return path;
-    }
-
-    [return: NotNullIfNotNull(nameof(path))]
-    private string? NormalizeDirectory(string? path) => string.IsNullOrEmpty(path) ? path : TrimAllEndingDirectorySeparators(this.pathUtilityService.NormalizePath(path));
-
-    /// <summary>
-    /// Given a path under sourceDirectory, and the same path in another filesystem,
-    /// determine what path could be replaced with sourceDirectory.
-    /// </summary>
-    /// <param name="sourceDirectoryBasedPath">Some directory path under sourceDirectory, including sourceDirectory.</param>
-    /// <param name="rebasePath">Path to the same directory as <paramref name="sourceDirectoryBasedPath"/> but in a different root.</param>
-    /// <returns>Portion of <paramref name="rebasePath"/> that corresponds to root, or null if it can not be rebased.</returns>
-    private string? GetRootRebasePath(string sourceDirectoryBasedPath, string? rebasePath)
-    {
-        if (string.IsNullOrEmpty(rebasePath) || string.IsNullOrEmpty(this.sourceDirectory) || string.IsNullOrEmpty(sourceDirectoryBasedPath))
-        {
-            return null;
-        }
-
-        // sourceDirectory is normalized, normalize others
-        sourceDirectoryBasedPath = this.NormalizeDirectory(sourceDirectoryBasedPath);
-        rebasePath = this.NormalizeDirectory(rebasePath);
-
-        // nothing to do if the paths are the same
-        if (rebasePath.Equals(sourceDirectoryBasedPath, StringComparison.Ordinal))
-        {
-            return null;
-        }
-
-        // find the relative path under sourceDirectory.
-        var sourceDirectoryRelativePath = this.NormalizeDirectory(Path.GetRelativePath(this.sourceDirectory!, sourceDirectoryBasedPath));
-
-        this.Logger.LogDebug("Attempting to rebase {RebasePath} to {SourceDirectoryBasedPath} using relative {SourceDirectoryRelativePath}", rebasePath, sourceDirectoryBasedPath, sourceDirectoryRelativePath);
-
-        // if the rebase path has the same relative portion, then we have a replacement.
-        if (rebasePath.EndsWith(sourceDirectoryRelativePath))
-        {
-            return rebasePath[..^sourceDirectoryRelativePath.Length];
-        }
-
-        // The path didn't have a common relative path, it might have been copied from a completely different location since it was built.
-        // We cannot rebase the paths.
-        return null;
-    }
-
-    private async Task<string?> RunDotNetVersionAsync(string workingDirectoryPath, CancellationToken cancellationToken)
-    {
-        var workingDirectory = new DirectoryInfo(workingDirectoryPath);
-
-        try
-        {
-            var process = await this.commandLineInvocationService.ExecuteCommandAsync("dotnet", ["dotnet.exe"], workingDirectory, cancellationToken, "--version").ConfigureAwait(false);
-
-            if (process.ExitCode != 0)
-            {
-                // debug only - it could be that dotnet is not actually on the path and specified directly by the build scripts.
-                this.Logger.LogDebug("Failed to invoke 'dotnet --version'. Return: {Return} StdErr: {StdErr} StdOut: {StdOut}.", process.ExitCode, process.StdErr, process.StdOut);
-                return null;
-            }
-
-            return process.StdOut.Trim();
-        }
-        catch (InvalidOperationException ioe)
-        {
-            // debug only - it could be that dotnet is not actually on the path and specified directly by the build scripts.
-            this.Logger.LogDebug("Failed to invoke 'dotnet --version'. {Message}", ioe.Message);
-            return null;
-        }
-    }
-
     public override Task<IndividualDetectorScanResult> ExecuteDetectorAsync(ScanRequest request, CancellationToken cancellationToken = default)
     {
-        this.sourceDirectory = this.NormalizeDirectory(request.SourceDirectory.FullName);
-        this.sourceFileRootDirectory = this.NormalizeDirectory(request.SourceFileRoot?.FullName);
-
+        this.projectInfoProvider.Initialize(request.SourceDirectory.FullName, request.SourceFileRoot?.FullName);
         return base.ExecuteDetectorAsync(request, cancellationToken);
     }
 
@@ -163,215 +63,10 @@ public class DotNetComponentDetector : FileComponentDetector
             return;
         }
 
-        var projectAssetsDirectory = this.pathUtilityService.GetParentDirectory(processRequest.ComponentStream.Location);
-        var projectPath = lockFile.PackageSpec.RestoreMetadata.ProjectPath;
-        var projectOutputPath = lockFile.PackageSpec.RestoreMetadata.OutputPath;
-
-        // The output path should match the location that the assets file, if it doesn't we could be analyzing paths
-        // on a different filesystem root than they were created.
-        // Attempt to rebase paths based on the difference between this file's location and the output path.
-        var rebasePath = this.GetRootRebasePath(projectAssetsDirectory, projectOutputPath);
-
-        if (rebasePath is not null)
-        {
-            projectPath = Path.Combine(this.sourceDirectory!, Path.GetRelativePath(rebasePath, projectPath));
-            projectOutputPath = Path.Combine(this.sourceDirectory!, Path.GetRelativePath(rebasePath, projectOutputPath));
-        }
-
-        if (!this.fileUtilityService.Exists(projectPath))
-        {
-            // Could be the assets file was not actually from this build
-            this.Logger.LogWarning("Project path {ProjectPath} specified by {ProjectAssetsPath} does not exist.", projectPath, processRequest.ComponentStream.Location);
-        }
-
-        var projectDirectory = this.pathUtilityService.GetParentDirectory(projectPath);
-        var sdkVersion = await this.GetSdkVersionAsync(projectDirectory, cancellationToken);
-
-        var projectName = lockFile.PackageSpec.RestoreMetadata.ProjectName;
-
-        if (!this.directoryUtilityService.Exists(projectOutputPath))
-        {
-            this.Logger.LogWarning("Project output path {ProjectOutputPath} specified by {ProjectAssetsPath} does not exist.", projectOutputPath, processRequest.ComponentStream.Location);
-
-            // default to use the location of the assets file.
-            projectOutputPath = projectAssetsDirectory;
-        }
-
-        var targetType = this.GetProjectType(projectOutputPath, projectName, cancellationToken);
-
-        var componentReporter = this.ComponentRecorder.CreateSingleFileComponentRecorder(projectPath);
-        foreach (var target in lockFile.Targets ?? [])
-        {
-            var targetFramework = target.TargetFramework;
-            var isSelfContained = this.IsSelfContained(lockFile.PackageSpec, targetFramework, target);
-            var targetTypeWithSelfContained = this.GetTargetTypeWithSelfContained(targetType, isSelfContained);
-
-            componentReporter.RegisterUsage(new DetectedComponent(new DotNetComponent(sdkVersion, targetFramework?.GetShortFolderName(), targetTypeWithSelfContained)));
-        }
-    }
-
-    private string? GetProjectType(string projectOutputPath, string projectName, CancellationToken cancellationToken)
-    {
-        if (this.directoryUtilityService.Exists(projectOutputPath) &&
-            projectName is not null &&
-            projectName.IndexOfAny(Path.GetInvalidFileNameChars()) == -1)
-        {
-            try
-            {
-                // look for the compiled output, first as dll then as exe.
-                var candidates = this.directoryUtilityService.EnumerateFiles(projectOutputPath, projectName + ".dll", SearchOption.AllDirectories)
-                        .Concat(this.directoryUtilityService.EnumerateFiles(projectOutputPath, projectName + ".exe", SearchOption.AllDirectories));
-                foreach (var candidate in candidates)
-                {
-                    try
-                    {
-                        return this.IsApplication(candidate) ? "application" : "library";
-                    }
-                    catch (Exception e)
-                    {
-                        this.Logger.LogWarning("Failed to open output assembly {AssemblyPath} error {Message}.", candidate, e.Message);
-                    }
-                }
-            }
-            catch (IOException e)
-            {
-                this.Logger.LogWarning("Failed to enumerate output directory {OutputPath} error {Message}.", projectOutputPath, e.Message);
-            }
-        }
-
-        return null;
-    }
-
-    private bool IsApplication(string assemblyPath)
-    {
-        using var peReader = new PEReader(this.fileUtilityService.MakeFileStream(assemblyPath));
-
-        // despite the name `IsExe` this is actually based of the CoffHeader Characteristics
-        return peReader.PEHeaders.IsExe;
-    }
-
-    private bool IsSelfContained(PackageSpec packageSpec, NuGetFramework? targetFramework, LockFileTarget target)
-    {
-        // PublishAot projects reference Microsoft.DotNet.ILCompiler, which implies
-        // native AOT compilation and therefore a self-contained deployment.
-        if (target?.Libraries != null &&
-            target.Libraries.Any(lib => "Microsoft.DotNet.ILCompiler".Equals(lib.Name, StringComparison.OrdinalIgnoreCase)))
-        {
-            return true;
-        }
-
-        if (packageSpec?.TargetFrameworks == null || targetFramework == null)
-        {
-            return false;
-        }
-
-        var targetFrameworkInfo = packageSpec.TargetFrameworks.FirstOrDefault(tf => tf.FrameworkName == targetFramework);
-        if (targetFrameworkInfo == null)
-        {
-            return false;
-        }
-
-        var frameworkReferences = targetFrameworkInfo.FrameworkReferences;
-        var packageDownloads = targetFrameworkInfo.DownloadDependencies;
-
-        if (frameworkReferences == null || frameworkReferences.Count == 0 || packageDownloads.IsDefaultOrEmpty)
-        {
-            return false;
-        }
-
-        foreach (var frameworkRef in frameworkReferences)
-        {
-            var frameworkName = frameworkRef.Name;
-            var hasRuntimeDownload = packageDownloads.Any(pd => pd.Name.StartsWith($"{frameworkName}.Runtime", StringComparison.OrdinalIgnoreCase));
-
-            if (hasRuntimeDownload)
-            {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private string? GetTargetTypeWithSelfContained(string? targetType, bool isSelfContained)
-    {
-        if (string.IsNullOrWhiteSpace(targetType))
-        {
-            return targetType;
-        }
-
-        return isSelfContained ? $"{targetType}-selfcontained" : targetType;
-    }
-
-    /// <summary>
-    /// Recursively get the sdk version from the project directory or parent directories.
-    /// </summary>
-    /// <param name="projectDirectory">Directory to start the search.</param>
-    /// <param name="cancellationToken">Cancellation token to halt the search.</param>
-    /// <returns>Sdk version found, or null if no version can be detected.</returns>
-    private async Task<string?> GetSdkVersionAsync(string? projectDirectory, CancellationToken cancellationToken)
-    {
-        // normalize since we need to use as a key
-        projectDirectory = this.NormalizeDirectory(projectDirectory);
-
-        if (string.IsNullOrWhiteSpace(projectDirectory))
-        {
-            // not expected
-            return null;
-        }
-
-        if (this.sdkVersionCache.TryGetValue(projectDirectory, out var sdkVersion))
-        {
-            return sdkVersion;
-        }
-
-        var parentDirectory = this.pathUtilityService.GetParentDirectory(projectDirectory);
-        var globalJsonPath = Path.Combine(projectDirectory, GlobalJsonFileName);
-
-        if (this.fileUtilityService.Exists(globalJsonPath))
-        {
-            sdkVersion = await this.RunDotNetVersionAsync(projectDirectory, cancellationToken);
-
-            if (string.IsNullOrWhiteSpace(sdkVersion))
-            {
-                var globalJson = await JsonDocument.ParseAsync(this.fileUtilityService.MakeFileStream(globalJsonPath), cancellationToken: cancellationToken, options: this.jsonDocumentOptions).ConfigureAwait(false);
-                if (globalJson.RootElement.TryGetProperty("sdk", out var sdk))
-                {
-                    if (sdk.TryGetProperty("version", out var version))
-                    {
-                        sdkVersion = version.GetString();
-                    }
-                }
-            }
-
-            if (!string.IsNullOrWhiteSpace(sdkVersion))
-            {
-                var globalJsonComponent = new DetectedComponent(new DotNetComponent(sdkVersion));
-                var recorder = this.ComponentRecorder.CreateSingleFileComponentRecorder(globalJsonPath);
-                recorder.RegisterUsage(globalJsonComponent, isExplicitReferencedDependency: true);
-                return sdkVersion;
-            }
-
-            // global.json may be malformed, or the sdk version may not be specified.
-        }
-
-        if (projectDirectory.Equals(this.sourceDirectory, StringComparison.OrdinalIgnoreCase) ||
-            projectDirectory.Equals(this.sourceFileRootDirectory, StringComparison.OrdinalIgnoreCase) ||
-            string.IsNullOrEmpty(parentDirectory) ||
-            projectDirectory.Equals(parentDirectory, StringComparison.OrdinalIgnoreCase))
-        {
-            // if we are at the source directory, source file root, or have reached a root directory, run `dotnet --version`
-            // this could fail if dotnet is not on the path, or if the global.json is malformed
-            sdkVersion = await this.RunDotNetVersionAsync(projectDirectory, cancellationToken);
-        }
-        else
-        {
-            // recurse up the directory tree
-            sdkVersion = await this.GetSdkVersionAsync(parentDirectory, cancellationToken);
-        }
-
-        this.sdkVersionCache[projectDirectory] = sdkVersion;
-
-        return sdkVersion;
+        await this.projectInfoProvider.RegisterDotNetComponentsAsync(
+            lockFile,
+            processRequest.ComponentStream.Location,
+            this.ComponentRecorder,
+            cancellationToken);
     }
 }
